+++ SimpleSDMLayers/ext/MakieExtension.jl
@@ -1,6 +1,6 @@
 module MakieExtension
 
-using MakieCore
+import Makie
 using SimpleSDMLayers
 
 function _sprinkle(layer::SDMLayer)
@@ -13,19 +13,19 @@ function _sprinkle(layer::SDMLayer)
     )
 end
 
-function MakieCore.convert_arguments(::MakieCore.GridBased, layer::SDMLayer)
+function Makie.convert_arguments(::Makie.GridBased, layer::SDMLayer)
     return _sprinkle(convert(SDMLayer{Float32}, layer))
 end
 
-MakieCore.convert_arguments(P::MakieCore.NoConversion, layer::SDMLayer) = MakieCore.convert_arguments(P, values(layer))
+Makie.convert_arguments(P::Makie.NoConversion, layer::SDMLayer) = Makie.convert_arguments(P, values(layer))
 
-function MakieCore.convert_arguments( P::MakieCore.PointBased, layer1::T1, layer2::T2, ) where {T1 <: SDMLayer, T2 <: SDMLayer}
+function Makie.convert_arguments( P::Makie.PointBased, layer1::T1, layer2::T2, ) where {T1 <: SDMLayer, T2 <: SDMLayer}
     k = findall(layer1.indices .& layer2.indices)
-    return MakieCore.convert_arguments(P, layer1[k], layer2[k])
+    return Makie.convert_arguments(P, layer1[k], layer2[k])
 end
 
-function MakieCore.convert_arguments( P::MakieCore.PointBased, layer::T, ) where {T <: SDMLayer{Bool}}
-    return MakieCore.convert_arguments(P, _sprinkle(ones(nodata(layer, false), Float32))...)
+function Makie.convert_arguments( P::Makie.PointBased, layer::T, ) where {T <: SDMLayer{Bool}}
+    return Makie.convert_arguments(P, _sprinkle(ones(nodata(layer, false), Float32))...)
 end
 
 end